feat(dataset): dataset adoption endpoints
