+++ app/api/api.py
@@ -1,26 +1,11 @@
-import os
-import httpx
-from dotenv import load_dotenv
-from typing import Optional, Literal
-
-from fastapi import (
-    APIRouter,
-    Query,
-    HTTPException
-    )
-
-from app.api.schemas.dataset import DatasetInfo, LocalDatasetImportResponse
+from fastapi import APIRouter
 from app.api.endpoints import (
+    dataset,
     histories,
     tools,
     workflows,
     invocation
     )
-
-from app.context import current_api_key
-from app.config import GALAXY_URL
-
-load_dotenv()
 
 api_router = APIRouter()
 
@@ -51,57 +36,9 @@ api_router.include_router(
     tags=["Invocation"]
 )
 
-
-@api_router.post(
-"/datasets/local_import",
-response_model = LocalDatasetImportResponse,
-summary="Import/Adopt a local dataset into galaxy, return the existing history_id and the dataset_id of the adopted file.",
-tags = ["Histories & Data"]
-
+# Include the dataset router with a prefix and tags
+api_router.include_router(
+    dataset.router,
+    prefix="/dataset",
+    tags=["Date Adoption"]
 )
-async def proxy_adopt_local_file(
-    file_path: str = Query(..., description="Absolute path to the local file"),
-    file_name: Optional[str] = Query(None),
-    extension: Optional[str] = Query(None),
-    file_origin: Optional[Literal["annotation", "hypothesis"]] = Query(None),
-    hist_id: Optional[str] = Query(None),
-    ):
-    
-    params = {
-    "file_path": file_path,
-    "file_name": file_name,
-    "extension": extension,
-    "file_origin": file_origin,
-    "hist_id": hist_id,
-    }
-
-
-    # remove None values to avoid polluting query string
-    params = {k: v for k, v in params.items() if v is not None}
-    api_key = current_api_key.get()
-    headers = {"x-api-key": api_key}
-    
-
-    async with httpx.AsyncClient(timeout=10) as client:
-        try:
-            
-            resp = await client.post(
-                url = f"{GALAXY_URL}/api/datasets/adopt_file",
-                params=params,
-                headers=headers,
-            )
-            
-        except httpx.RequestError as exc:
-            raise HTTPException(
-                status_code=502,
-                detail=f"Upstream service unreachable: {exc}",
-                )
-
-    if resp.status_code >= 400:
-        raise HTTPException(
-            status_code=resp.status_code,
-            detail=resp.text,
-            )
-
-    response = resp.json()
-    return LocalDatasetImportResponse(**response)
